+++ packages/core/src/repository/data-source/data-source.ts
@@ -2,18 +2,10 @@ import { Query } from '..';
 
 export interface GetDataSource<T> {
     get: (query: Query) => Promise<T>;
-    /**
-     * @deprecated please use get with an array type instead
-     */
-    getAll: (query: Query) => Promise<T[]>;
 }
 
 export interface PutDataSource<T> {
     put: (value: T | undefined, query: Query) => Promise<T>;
-    /**
-     * @deprecated please use put with an array type instead
-     */
-    putAll: (values: T[] | undefined, query: Query) => Promise<T[]>;
 }
 
 export interface DeleteDataSource {
