+++ packages/core/src/repository/index.ts
@@ -17,12 +17,11 @@ export * from './void.repository';
 export * from './cache.repository';
 export * from './interactor/delete.interactor';
 export * from './interactor/get.interactor';
-export * from './interactor/get-all.interactor';
 export * from './interactor/put.interactor';
-export * from './interactor/put-all.interactor';
-export * from './interactor/put-all.interactor';
 export * from './data-source/sql/sql.query';
 export * from './data-source/sql/sql.constants';
 export * from './data-source/sql/sql-row-counter.data-source';
 export * from './data-source/sql/raw-sql.data-source';
+export * from './data-source/sql/abstract-raw-sql.data-source';
+export * from './data-source/sql/array-raw-sql.data-source';
 export * from './data-source/sql/sql-query-param-composer';
