+++ packages/core/src/repository/repository-mapper.ts
@@ -33,24 +33,8 @@ export class RepositoryMapper<In, Out> implements Repository<Out> {
         return this.getMapper.get(query, operation);
     }
 
-    /**
-     * @deprecated please use get with an array type instead
-     */
-    public getAll(query: Query, operation: Operation): Promise<Out[]> {
-        console.warn('getAll is deprecated. Please use get instead');
-        return this.getMapper.getAll(query, operation);
-    }
-
     public put(value: Out | undefined, query: Query, operation: Operation): Promise<Out> {
         return this.putMapper.put(value, query, operation);
-    }
-
-    /**
-     * @deprecated please use put with an array type instead
-     */
-    public async putAll(values: Out[] | undefined, query: Query, operation: Operation): Promise<Out[]> {
-        console.warn('putAll is deprecated. Please use put instead');
-        return this.putMapper.putAll(values, query, operation);
     }
 
     public async delete(query: Query, operation: Operation): Promise<void> {
@@ -70,15 +54,6 @@ export class GetRepositoryMapper<In, Out> implements GetRepository<Out> {
     public async get(query: Query, operation: Operation): Promise<Out> {
         const result: In = await this.getRepository.get(query, operation);
         return this.toOutMapper.map(result);
-    }
-
-    /**
-     * @deprecated please use get with an array type instead
-     */
-    public async getAll(query: Query, operation: Operation): Promise<Out[]> {
-        console.warn('getAll is deprecated. Please use get instead');
-        const results: In[] = await this.getRepository.getAll(query, operation);
-        return results.map((r: In) => this.toOutMapper.map(r));
     }
 }
 
@@ -100,15 +75,5 @@ export class PutRepositoryMapper<In, Out> implements PutRepository<Out> {
         const mapped: In | undefined = value ? this.toInMapper.map(value) : undefined;
         const result: In = await this.putRepository.put(mapped, query, operation);
         return this.toOutMapper.map(result);
-    }
-
-    /**
-     * @deprecated please use put with an array type instead
-     */
-    public async putAll(values: Out[] | undefined, query: Query, operation: Operation): Promise<Out[]> {
-        console.warn('putAll is deprecated. Please use put instead');
-        const mapped: In[] | undefined = values ? values.map((v) => this.toInMapper.map(v)) : undefined;
-        const results: In[] = await this.putRepository.putAll(mapped, query, operation);
-        return results.map((r: In) => this.toOutMapper.map(r));
     }
 }
