+++ packages/core/src/repository/single-data-source.repository.ts
@@ -2,38 +2,20 @@ import { DeleteDataSource, GetDataSource, PutDataSource } from './data-source/da
 import { Operation } from './operation/operation';
 import { Query } from './query/query';
 import { DeleteRepository, GetRepository, PutRepository, Repository } from './repository';
-import { DeviceConsoleLogger, Logger } from '../helpers';
 
 export class SingleDataSourceRepository<T> implements Repository<T> {
     constructor(
         private readonly getDataSource: GetDataSource<T>,
         private readonly putDataSource: PutDataSource<T>,
         private readonly deleteDataSource: DeleteDataSource,
-        private readonly logger: Logger = new DeviceConsoleLogger(),
     ) {}
 
     public get(query: Query, _operation: Operation): Promise<T> {
         return this.getDataSource.get(query);
     }
 
-    /**
-     * @deprecated please use get with an array type instead
-     */
-    public getAll(query: Query, _operation: Operation): Promise<T[]> {
-        console.warn('getAll is deprecated. Please use get instead');
-        return this.getDataSource.getAll(query);
-    }
-
     public put(value: T | undefined, query: Query, _operation: Operation): Promise<T> {
         return this.putDataSource.put(value, query);
-    }
-
-    /**
-     * @deprecated please use put with an array type instead
-     */
-    public putAll(values: T[] | undefined, query: Query, _operation: Operation): Promise<T[]> {
-        console.warn('putAll is deprecated. Please use put instead');
-        return this.putDataSource.putAll(values, query);
     }
 
     public delete(query: Query, _operation: Operation): Promise<void> {
@@ -47,14 +29,6 @@ export class SingleGetDataSourceRepository<T> implements GetRepository<T> {
     public get(query: Query, _operation: Operation): Promise<T> {
         return this.getDataSource.get(query);
     }
-
-    /**
-     * @deprecated please use get with an array type instead
-     */
-    public getAll(query: Query, _operation: Operation): Promise<T[]> {
-        console.warn('getAll is deprecated. Please use get instead');
-        return this.getDataSource.getAll(query);
-    }
 }
 
 export class SinglePutDataSourceRepository<T> implements PutRepository<T> {
@@ -63,21 +37,10 @@ export class SinglePutDataSourceRepository<T> implements PutRepository<T> {
     public put(value: T | undefined, query: Query, _operation: Operation): Promise<T> {
         return this.putDataSource.put(value, query);
     }
-
-    /**
-     * @deprecated please use put with an array type instead
-     */
-    public putAll(values: T[] | undefined, query: Query, _operation: Operation): Promise<T[]> {
-        console.warn('putAll is deprecated. Please use put instead');
-        return this.putDataSource.putAll(values, query);
-    }
 }
 
 export class SingleDeleteDataSourceRepository implements DeleteRepository {
-    constructor(
-        private readonly deleteDataSource: DeleteDataSource,
-        private readonly logger: Logger = new DeviceConsoleLogger(),
-    ) {}
+    constructor(private readonly deleteDataSource: DeleteDataSource) {}
 
     public delete(query: Query, _operation: Operation): Promise<void> {
         return this.deleteDataSource.delete(query);
