+++ packages/core/src/repository/void.repository.ts
@@ -5,28 +5,14 @@ import { DeleteRepository, GetRepository, PutRepository, Repository } from './re
 import { DeviceConsoleLogger, Logger } from '../helpers';
 
 export class VoidRepository<T> implements Repository<T> {
-    constructor(private readonly logger: Logger = new DeviceConsoleLogger()) {}
-
     public async get(_query: Query, _operation: Operation): Promise<T> {
         throw new MethodNotImplementedError('Called get on VoidRepository');
     }
-    /**
-     * @deprecated please use get with an array type instead
-     */
-    public async getAll(_query: Query, _operation: Operation): Promise<T[]> {
-        console.warn('getAll is deprecated. Please use get instead');
-        throw new MethodNotImplementedError('Called getAll on VoidRepository');
-    }
+
     public async put(_value: T | undefined, _query?: Query, _operation?: Operation): Promise<T> {
         throw new MethodNotImplementedError('Called put on VoidRepository');
     }
-    /**
-     * @deprecated please use put with an array type instead
-     */
-    public async putAll(_values: T[] | undefined, _query?: Query, _operation?: Operation): Promise<T[]> {
-        console.warn('putAll is deprecated. Please use put instead');
-        throw new MethodNotImplementedError('Called putAll on VoidRepository');
-    }
+
     public async delete(_query: Query, _operation: Operation): Promise<void> {
         throw new MethodNotImplementedError('Called delete on VoidRepository');
     }
@@ -36,26 +22,11 @@ export class VoidGetRepository<T> implements GetRepository<T> {
     public async get(_query: Query, _operation: Operation): Promise<T> {
         throw new MethodNotImplementedError('Called get on VoidGetRepository');
     }
-
-    /**
-     * @deprecated please use get with an array type instead
-     */
-    public async getAll(_query: Query, _operation: Operation): Promise<T[]> {
-        console.warn('getAll is deprecated. Please use get instead');
-        throw new MethodNotImplementedError('Called getAll on VoidGetRepository');
-    }
 }
 
 export class VoidPutRepository<T> implements PutRepository<T> {
     public async put(_value: T | undefined, _query?: Query, _operation?: Operation): Promise<T> {
         throw new MethodNotImplementedError('Called put on VoidPutRepository');
-    }
-    /**
-     * @deprecated please use put with an array type instead
-     */
-    public async putAll(_values: T[] | undefined, _query?: Query, _operation?: Operation): Promise<T[]> {
-        console.warn('putAll is deprecated. Please use put instead');
-        throw new MethodNotImplementedError('Called putAll on VoidPutRepository');
     }
 }
